refactor(T10123): extract SAGA constants + resolveSagaMemberIds
Move SAGA_LABEL, SAGA_GROUPS_RELATION, LIST_BINDING_SAGA_GROUPS, and the
resolveSagaMemberIds helper out of packages/core/src/tasks/list.ts into a
dedicated packages/core/src/sagas/ module (constants.ts + storage.ts +
index.ts barrel).

list.ts re-exports the three constants for backwards-compat with existing
test fixtures and external consumers — new code should import from
@cleocode/core/sagas. listTasks() itself now calls the moved helper
directly (no compat shim).

Subtask of T10120 (A6.1). Saga T10113. Epic T10208.
